Add support for ANY psr7 facotory
