+++ composer.json
@@ -18,15 +18,17 @@
     "require": {
         "php": "^7.1",
         "psr/http-message": "^1.0",
+        "psr/http-factory": "^1.0",
+        "nyholm/psr7-server": "^0.2.1",
         "symfony/http-foundation": "^3.4 || ^4.0"
     },
     "require-dev": {
         "symfony/phpunit-bridge": "^3.4 || 4.0",
-        "zendframework/zend-diactoros": "^1.8"
+        "http-interop/http-factory-diactoros": "^1.0"
     },
     "suggest": {
         "psr/http-message-implementation": "To use the HttpFoundation factory",
-        "zendframework/zend-diactoros": "To use the Zend Diactoros factory"
+        "psr/http-factory-implementation": "To use the PSR7factory"
     },
     "autoload": {
         "psr-4": { "Symfony\\Bridge\\PsrHttpMessage\\": "" }
